refactor: 카드 추가 폼 리팩토링
