--- src/components/SecurityCodeGuide/index.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import securityGuideImage from '../../assets/securityGuideImage.svg';
-
-export default function SecurityCodeGuide() {
-  return <img src={securityGuideImage} alt="securityGuideImage" width="280px"></img>;
-}
