--- src/components/SecurityCodeGuide/stories.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import SecurityCodeGuide from '.';
-
-export default {
-  title: 'Components/SecurityCodeGuide',
-  component: SecurityCodeGuide,
-};
-
-const Template = (args) => <SecurityCodeGuide {...args} />;
-
-export const Default = Template.bind({});
-
-Default.args = {};
