+++ src/pages/AddCardForm/index.js
@@ -1,24 +1,23 @@
 import React, { useState, useRef, useEffect, useContext } from 'react';
-import { CARD, CARD_COMPANY, ERROR_MESSAGE, PATH } from '../../constants';
+import { CARD, CARD_COMPANY, PATH } from '../../constants';
 import {
   Icon,
   Card,
   Input,
   Header,
   TextButton,
-  InputWithCounter,
   CardPasswordInput,
   Modal,
   CardCompanySelection,
-  SecurityCodeGuide,
 } from '../../components';
-import { cardSerialNumberFormatter, MMYYDateFormatter } from '../../utils/formatter';
-import { isValidSerialNumber, isValidDateFormat, isValidUserName } from './validator';
-import './style.css';
 import { useHistory } from 'react-router-dom';
 import { CardsContext } from '../../cardsContext';
 import { idGenerator } from '../../utils/idGenerator';
 import useForm from '../../hooks/useForm';
+import SerialNumberInput from './Inputs/SerialNumberInput';
+import ExpirationDateInput from './Inputs/ExpirationDateInput';
+import UserNameInput from './Inputs/UserNameInput';
+import { ReactComponent as SecurityCodeGuide } from '../../assets/securityGuideImage.svg';
 
 const initialAddCardForm = {
   number: '',
@@ -29,26 +28,9 @@ const initialAddCardForm = {
   password: { first: '', second: '' },
 };
 
-export default function AddCardForm({ cards }) {
-  const dispatch = useContext(CardsContext);
-
-  const [cardNumberErrorMessage, setCardNumberErrorMessage] = useState('');
-  const [expirationDateErrorMessage, setExpirationDateErrorMessage] = useState('');
-  const [userNameErrorMessage, setUserNameErrorMessage] = useState('');
-  const [securityCodeErrorMessage, setSecurityCodeErrorMessage] = useState('');
-
-  const serialNumberInputElement = useRef(null);
-
+export default function AddCardForm() {
   const history = useHistory();
-
-  const [isModalOpened, setIsModalOpened] = useState(false);
-  const [modalContents, setModalContents] = useState('cardSelection');
-
-  const onSetModalContents = (name) => {
-    setModalContents(name);
-
-    setIsModalOpened(true);
-  };
+  const dispatch = useContext(CardsContext);
 
   const [
     { number, company, expirationDate, userName, securityCode, password },
@@ -57,16 +39,29 @@ export default function AddCardForm({ cards }) {
     reset,
   ] = useForm(initialAddCardForm);
 
+  const [securityCodeErrorMessage, setSecurityCodeErrorMessage] = useState('');
+
+  const serialInputElement = useRef(null);
+
+  const [isModalOpened, setIsModalOpened] = useState(false);
+  const [modalContents, setModalContents] = useState('cardSelection');
+
   const isFormCompleted =
-    isValidSerialNumber(number) &&
+    number.length === CARD.SERIAL_NUMBER_LENGTH &&
     company &&
-    isValidDateFormat(expirationDate) &&
+    expirationDate.length === CARD.EXPIRATION_DATE_LENGTH &&
     securityCode.length === CARD.SECURITY_CODE_LENGTH &&
     password.first &&
     password.second;
 
+  const onSetModalContents = (name) => {
+    setModalContents(name);
+
+    setIsModalOpened(true);
+  };
+
   useEffect(() => {
-    serialNumberInputElement.current.focus();
+    serialInputElement.current.focus();
   }, [company]);
 
   const onSetPassword = (key, event) => {
@@ -77,94 +72,12 @@ export default function AddCardForm({ cards }) {
     setInput(name, { ...password, [key]: value });
   };
 
-  const offsetByInputType = {
-    deleteContentBackward: -1,
-    insertText: 1,
-  };
-
-  const getOffset = (inputType, selectionStart) => {
-    if (inputType === 'insertText') {
-      return selectionStart !== 0 && selectionStart % (CARD.SERIAL_NUMBER_UNIT_LENGTH + 1) === 0
-        ? offsetByInputType[inputType]
-        : 0;
-    }
-
-    if (inputType === 'deleteContentBackward') {
-      return selectionStart !== 0 &&
-        (selectionStart + 1) % (CARD.SERIAL_NUMBER_UNIT_LENGTH + 1) === 0
-        ? offsetByInputType[inputType]
-        : 0;
-    }
-
-    return 0;
-  };
-
-  const getCurrentSerialNumber = {
-    deleteContentBackward: (serialIndex) => {
-      return number.slice(0, serialIndex - 1) + number.slice(serialIndex);
-    },
-    insertText: (serialIndex, insertKey) => {
-      return number.slice(0, serialIndex) + insertKey + number.slice(serialIndex);
-    },
-    insertFromPaste: () => {
-      throw new Error('붙여 넣기는 할 수 없습니다.');
-    },
-  };
-
-  const onSerialNumberInputChange = (event) => {
-    const inputKey = event.nativeEvent.data;
-    const inputValue = event.target.value.replaceAll('-', '');
-
-    if (isNaN(inputKey)) {
-      event.target.value = cardSerialNumberFormatter(number);
-      return;
-    }
-
-    const inputType = event.nativeEvent.inputType;
-    const selectionStart = event.target.selectionStart;
-    const offset = getOffset(inputType, selectionStart);
-    const currentLocation = selectionStart + offset;
-
-    const name = event.target.name;
-
-    try {
-      const serialIndex =
-        currentLocation -
-        Math.floor(currentLocation / (CARD.SERIAL_NUMBER_UNIT_LENGTH + 1)) -
-        offsetByInputType[inputType];
-      const currentSerialNumber = getCurrentSerialNumber[inputType](serialIndex, inputKey);
-
-      const value = cardSerialNumberFormatter(currentSerialNumber);
-
-      event.target.value = value;
-
-      setInput(name, currentSerialNumber);
-      setCardNumberErrorMessage('');
-
-      event.target.setSelectionRange(currentLocation, currentLocation);
-
-      if (inputValue.length === CARD.SERIAL_ID_CODE_LENGTH) {
-        serialNumberInputElement.current.blur();
-        onSetModalContents('cardSelection');
-      }
-
-      if (company && inputValue.length < CARD.SERIAL_ID_CODE_LENGTH) {
-        setInput(name, '');
-      }
-    } catch (error) {
-      setInput(name, '');
-      event.target.value = '';
-      setCardNumberErrorMessage(error.message);
-    }
-  };
-
   const onCardFormSubmit = (event) => {
     event.preventDefault();
 
     if (!isFormCompleted) return;
 
     const newCardId = idGenerator.getId();
-
     const newCard = {
       id: newCardId,
       userName: userName.toUpperCase(),
@@ -177,22 +90,24 @@ export default function AddCardForm({ cards }) {
     };
 
     dispatch({ type: 'ADD_CARD', card: newCard });
-
     reset();
 
     history.push(`${PATH.EDIT_CARD_NICKNAME}/${newCardId}`);
   };
 
   return (
-    <div className="add-card-form__container">
+    <div>
       <Header>
-        <button type="button">
+        <button type="button" onClick={() => history.goBack()}>
           <Icon.LeftArrow size="16px" color="#525252" />
         </button>
         <h1>카드 추가</h1>
       </Header>
-      <form className="add-card-form" onSubmit={onCardFormSubmit}>
-        <div className="card-preview">
+      <form
+        className="contents add-card-form w-100 d-flex flex-col relative items-center"
+        onSubmit={onCardFormSubmit}
+      >
+        <div className="mb-10">
           <Card
             userName={userName}
             companyName={CARD_COMPANY[company]?.NAME}
@@ -202,90 +117,15 @@ export default function AddCardForm({ cards }) {
           />
         </div>
 
-        <Input
-          id="card-number"
-          name="number"
-          type="tel"
-          label="카드번호"
-          inputStyle={{ width: '100%' }}
-          maxLength="19"
-          onChange={onSerialNumberInputChange}
-          onFocus={() => {
-            if (!company && number.length === CARD.SERIAL_ID_CODE_LENGTH) {
-              serialNumberInputElement.current.blur();
-              onSetModalContents('cardSelection');
-            }
-          }}
-          forwardRef={serialNumberInputElement}
-          inputmode="numeric"
-          textAlign="center"
-          errorMessage={cardNumberErrorMessage}
+        <SerialNumberInput
+          number={number}
+          company={company}
+          setInput={setInput}
+          onSetModalContents={onSetModalContents}
+          forwardRef={serialInputElement}
         />
-
-        <Input
-          id="expiration-date"
-          name="expirationDate"
-          type="text"
-          label="만료일"
-          inputStyle={{ width: '7rem' }}
-          placeholder="MM/YY"
-          textAlign="center"
-          maxLength="5"
-          inputmode="numeric"
-          value={expirationDate}
-          onChange={(event) => {
-            const inputValue = event.target.value.replace('/', '');
-            const { name } = event.target;
-
-            try {
-              if (isNaN(inputValue)) throw Error('숫자만 입력가능합니다.');
-
-              const formattedDate = MMYYDateFormatter(inputValue);
-              setInput(name, formattedDate);
-
-              if (formattedDate.length === CARD.EXPIRATION_DATE_LENGTH) {
-                if (isValidDateFormat(formattedDate)) {
-                  setExpirationDateErrorMessage('');
-                  return;
-                }
-
-                setExpirationDateErrorMessage(ERROR_MESSAGE.INVALID_DATE_FORMAT);
-              }
-            } catch (error) {
-              setInput(name, '');
-              setExpirationDateErrorMessage(error.message);
-            }
-          }}
-          onClick={({ target }) => {
-            setInput(target.name, '');
-            expirationDateErrorMessage && setExpirationDateErrorMessage('');
-          }}
-          errorMessage={expirationDateErrorMessage}
-        />
-
-        <InputWithCounter
-          id="user-name"
-          name="userName"
-          type="text"
-          inputStyle={{ width: '100%' }}
-          label="카드 소유자 이름(선택)"
-          placeholder="카드에 표시된 이름과 동일하게 입력하세요."
-          letterCounter={{ current: userName.length, max: CARD.USER_NAME_MAX_LENGTH }}
-          maxLength={CARD.USER_NAME_MAX_LENGTH}
-          value={userName}
-          onChange={(event) => {
-            const name = event.target.value;
-
-            if (!isValidUserName(name)) {
-              setUserNameErrorMessage(ERROR_MESSAGE.INVALID_USER_NAME);
-              return;
-            }
-
-            onInputChange(event);
-            setUserNameErrorMessage('');
-          }}
-          errorMessage={userNameErrorMessage}
-        />
+        <ExpirationDateInput expirationDate={expirationDate} setInput={setInput} />
+        <UserNameInput userName={userName} onInputChange={onInputChange} />
 
         <Input
           id="card-security-code"
@@ -312,7 +152,7 @@ export default function AddCardForm({ cards }) {
         >
           <button
             type="button"
-            className="security-code-guide-button"
+            className="ml-3"
             onClick={() => {
               onSetModalContents('questionMark');
             }}
