--- src/pages/AddCardForm/style.css
@@ -1,35 +1,0 @@
-.add-card-form__container {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  justify-content: center;
-  align-items: center;
-}
-
-.add-card-form {
-  width: 100%;
-  max-width: 720px;
-  padding: 0 28px 56px;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  position: relative;
-  flex-grow: 1;
-}
-
-.card-preview {
-  align-self: center;
-  margin-bottom: 2rem;
-}
-
-.security-code-guide-button {
-  display: inherit;
-  background: none;
-  outline: 0;
-  border: 0;
-  cursor: pointer;
-  margin: 0 0 0 12px;
-  padding: 0;
-  box-sizing: content-box;
-}
